A few formatting suggestions
I'm not attached to these, so reject or re-do if you don't like them. Additional suggestions: clarify language around options vs flags and commands vs programs. With an hour to cover this it will be primarily useful for review; it might be worth making a disclaimer for anyone with no experience that it will probably move a bit fast but that we have helpers to support review during work time.